+++ CHANGELOG.md
@@ -6,10 +6,12 @@
 - adding cross platform shared library loader
 - adding platform identifier
 - adding [imgui_widget_flamegraph](https://github.com/bwrsandman/imgui-flame-graph) and its profiler
-- adding a tile map renderer
+- adding a tilemap renderer
+- adding a tilemap example
 
 ### Changed
 - updating `include/Small/All.hpp` to include the shared library loader
+- updating `include/Small/All.hpp` to include the tilemap renderer
 
 ### Removed